requirements.txt
Flask==1.1.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
pyOpenSSL==20.0.1
idna==2.9
python-socketio==4.6.0
urllib3==1.25.8
validators==0.18.2
pem==21.1.0
service-identity==18.1.0
pytest==6.2.2
pytest-flask==1.2.0
prometheus-client==0.7.1
requests==2.23.0
ocsp-checker==1.6.0
